+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.5.0] - 2020-01-07
+### Added
+- Support for Azure Key Vault credentials (in `config.contrib`)
+- Tox support
+### Changed
+- Changed the `__repr__` and `__str__` methods so possibly sensitive values are not printed by default.
 
 ## [0.4.0] - 2019-10-11
 ### Added
@@ -27,8 +32,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Initial version
 
-[Unreleased]: https://github.com/tr11/python-configuration/compare/0.4.0...HEAD
-[0.3.1]: https://github.com/tr11/python-configuration/compare/0.3.1...0.4.0
+[Unreleased]: https://github.com/tr11/python-configuration/compare/0.5.0...HEAD
+[0.5.0]: https://github.com/tr11/python-configuration/compare/0.4.0...0.5.0
+[0.4.0]: https://github.com/tr11/python-configuration/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/tr11/python-configuration/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/tr11/python-configuration/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/tr11/python-configuration/compare/0.1.0...0.2.0